Final build of reservations is complete. No more buffer based on rate. It is much more stable but still not very efficient.

+++ instruction.cc
@@ -1,11 +1,11 @@
 #include "instruction.hh"
 
 
-instruction::instruction(int id, int inputSpeed, bool inputHasReservation){
+instruction::instruction(int id, float inputSpeed, bool inputHasReservation){
     this->id = id;
     this->speed = inputSpeed;
     this->hasReservation = inputHasReservation;
-    
+
 }
 
 void instruction::printid(){
